CDC-20 Refactor: RootLayout
