+++ apps/native/app/_layout.tsx
@@ -5,40 +5,36 @@ import '@/styles/unistyles';
 
 import { BottomSheetModalProvider } from '@gorhom/bottom-sheet';
 import { PortalProvider } from '@gorhom/portal';
-import NetInfo from '@react-native-community/netinfo';
 import { initializeKakaoSDK } from '@react-native-kakao/core';
 import { DarkTheme, DefaultTheme, ThemeProvider } from '@react-navigation/native';
-import {
-  onlineManager,
-  focusManager,
-  QueryCache,
-  QueryClient,
-  QueryClientProvider,
-} from '@tanstack/react-query';
+
 import { setDefaultOptions } from 'date-fns';
 import { ko } from 'date-fns/locale';
-import { useFonts } from 'expo-font';
-import { SplashScreen, Stack, useNavigationContainerRef } from 'expo-router';
-import { useEffect } from 'react';
+import { Stack } from 'expo-router';
+
+import { useCallback, useEffect } from 'react';
 import { AppState, LogBox, Platform } from 'react-native';
 import { GestureHandlerRootView } from 'react-native-gesture-handler';
 import 'react-native-reanimated';
 import { configureReanimatedLogger, ReanimatedLogLevel } from 'react-native-reanimated';
 import { SafeAreaProvider } from 'react-native-safe-area-context';
 import ErrorBoundary from 'react-native-error-boundary';
-import * as Notifications from 'expo-notifications';
-import { PushNotificationTrigger } from 'expo-notifications';
 
 import { useColorScheme } from '@/hooks/useColorScheme';
 import { useDevPlugins } from '@/hooks/useDevPlugins';
 import { useLoadUser } from '@/hooks/useLoadUser';
 import { useDiscountChannelsArrange } from '@/hooks/settings/useDiscountChannelArrange';
 import { useErrorHandler } from '@/hooks/useErrorHandler';
-import Sentry, { reactNavigationIntegration } from '@/libs/sentry';
+import Sentry from '@/libs/sentry';
 import { useExpoUpdate } from '@/hooks/useExpoUpdate';
 import CircularProgress from '@/components/core/progress/circular';
 import { usePushNotifications } from '@/hooks/notification/usePushNotifications';
 import StatusBar from '@/components/custom/status-bar';
+import { useSplashScreen } from '@/hooks/useSplashScreen';
+import { QueryClientProvider } from '@tanstack/react-query';
+import { queryClient, handleAppStateChange } from '@/libs/react-query/client';
+import { useAppState } from '@/hooks/useAppState';
+
 export { ErrorBoundary } from 'expo-router';
 
 LogBox.ignoreLogs(['Failed prop type']);
@@ -52,63 +48,8 @@ setDefaultOptions({ locale: ko });
 
 initializeKakaoSDK(process.env.EXPO_PUBLIC_KAKAO_TEST_NATIVE_APP_KEY ?? '');
 
-// Prevent the splash screen from auto-hiding before asset loading is complete.
-SplashScreen.preventAutoHideAsync();
-
-const queryClient = new QueryClient({
-  queryCache: new QueryCache({
-    onError: error => {
-      Sentry.captureException(error);
-    },
-  }),
-});
-
-onlineManager.setEventListener(setOnline => {
-  return NetInfo.addEventListener(state => {
-    setOnline(!!state.isConnected);
-  });
-});
-
-Notifications.setNotificationHandler({
-  handleNotification: async notification => {
-    //https://github.com/expo/expo/issues/31184
-    const trigger = notification.request.trigger as PushNotificationTrigger;
-    if (trigger?.type === 'push') {
-      const isDataOnly =
-        trigger?.remoteMessage?.notification === null ||
-        (trigger?.payload?.aps as any)['content-available'] === 1;
-
-      if (isDataOnly) {
-        return {
-          shouldShowAlert: false,
-          shouldPlaySound: false,
-          shouldSetBadge: false,
-        };
-      }
-    }
-    return {
-      shouldShowAlert: true,
-      shouldPlaySound: true,
-      shouldSetBadge: true,
-    };
-  },
-});
-
-function onAppStateChange(status: AppStateStatus) {
-  if (Platform.OS !== 'web') {
-    focusManager.setFocused(status === 'active');
-  }
-}
-
 function RootLayout() {
-  const navigationRef = useNavigationContainerRef();
-
-  useDevPlugins({ queryClient, navigationRef });
-
-  const [loaded] = useFonts({
-    SpaceMono: require('../assets/fonts/SpaceMono-Regular.ttf'),
-    Inter: require('../assets/fonts/Inter_18pt-Medium.ttf'),
-  });
+  useDevPlugins({ queryClient });
 
   const { isUpdating, checkUpdate } = useExpoUpdate();
 
@@ -122,28 +63,21 @@ function RootLayout() {
 
   const { theme } = useColorScheme(true);
 
-  useEffect(() => {
-    if (navigationRef) {
-      reactNavigationIntegration.registerNavigationContainer(navigationRef);
-    }
-  }, [navigationRef]);
+  const { appIsReady } = useSplashScreen();
 
-  useEffect(() => {
-    if (loaded) {
-      SplashScreen.hideAsync();
-    }
-  }, [loaded]);
+  useAppState(
+    useCallback(
+      (status: AppStateStatus) => {
+        handleAppStateChange(status);
+        if (status === 'active') {
+          checkUpdate();
+        }
+      },
+      [checkUpdate],
+    ),
+  );
 
-  useEffect(() => {
-    const subscription = AppState.addEventListener('change', status => {
-      onAppStateChange(status);
-      status === 'active' && checkUpdate();
-    });
-
-    return () => subscription.remove();
-  }, [checkUpdate]);
-
-  if (!loaded || isUpdating) {
+  if (!appIsReady || isUpdating) {
     return <CircularProgress style={{ flex: 1 }} />;
   }
