+++ apps/native/hooks/useSplashScreen.ts
@@ -1,0 +1,26 @@
+import { useFonts } from 'expo-font';
+import { useLayoutEffect } from 'react';
+import * as SplashScreen from 'expo-splash-screen';
+
+SplashScreen.setOptions({
+  duration: 1000,
+  fade: true,
+});
+
+// Prevent the splash screen from auto-hiding before asset loading is complete.
+SplashScreen.preventAutoHideAsync();
+
+export function useSplashScreen() {
+  const [appIsReady] = useFonts({
+    SpaceMono: require('../assets/fonts/SpaceMono-Regular.ttf'),
+    Inter: require('../assets/fonts/Inter_18pt-Medium.ttf'),
+  });
+
+  useLayoutEffect(() => {
+    if (appIsReady) {
+      SplashScreen.hideAsync();
+    }
+  }, [appIsReady]);
+
+  return { appIsReady };
+}
