+++ apps/native/package.json
@@ -43,7 +43,7 @@
     "@tanstack/react-query": "^5.62.3",
     "date-fns": "^4.1.0",
     "es-toolkit": "^1.30.1",
-    "expo": "52.0.18",
+    "expo": "52.0.19",
     "expo-apple-authentication": "~7.1.2",
     "expo-build-properties": "~0.13.1",
     "expo-constants": "~17.0.3",
@@ -53,7 +53,7 @@
     "expo-image": "2.0.3",
     "expo-linking": "~7.0.3",
     "expo-notifications": "~0.29.11",
-    "expo-router": "4.0.11",
+    "expo-router": "~4.0.13",
     "expo-splash-screen": "~0.29.18",
     "expo-status-bar": "~2.0.0",
     "expo-system-ui": "~4.0.6",
